+++ packages/terra-clinical-onset-picker/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Fixed
+  * Updated so that onOnsetChange gets called when a complete invalid date is entered.
+
 ## 4.31.0 - (November 16, 2021)
 
 * Changed